# services/mapi-xdg/docs/indexer/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dogecoin - Blockchain Indexer API",
    "description": "Dogecoin indexer endpoints: addresses, dunes, inscriptions and DRC-20. The same binary also proxies a set of node RPC endpoints (blocks, mempool, chain info) documented in the repository README.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.txt"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local instance"
    }
  ],
  "paths": {
...
    "/addresses/{address}/inscriptions": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "Inscriptions by Address",
        "description": "List of all inscriptions which reside at the specified address or script pubkey.",
        "operationId": "inscriptions_by_address",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Dogecoin address or hex encoded script pubkey",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "DNG3G7pKc1DgciyWs36GFUnVyvJieDhKdb"
          },
          {
            "name": "count",
            "in": "query",
            "description": "The max number of results per page",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "type": "integer",
                  "default": 100,
                  "minimum": 0
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor string, use the cursor included in a page of results to fetch the next page",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedInscriptionByAddress"
                },
                "example": {
                  "data": [
                    {
                      "inscription_id": "f02da3d6bebab13d5d604be1ed73d9a9c677dadf6ca71bc5fff7d99cdead11b0i0",
                      "satoshis": "546",
                      "utxo_block_height": 843010,
                      "utxo_confirmations": 23701,
                      "utxo_sat_offset": 0,
                      "utxo_txid": "e2283e7c915ef074806136e0002cbc69f5fdd2e9f70f14b0eab48cdcbe867cc1",
                      "utxo_vout": 0
                    },
                    {
                      "inscription_id": "7d0a2dd897222913d58fc957b0429526117a0a61c964642fe93b077f328ccec1i0",
                      "satoshis": "546",
                      "utxo_block_height": 850976,
                      "utxo_confirmations": 15735,
                      "utxo_sat_offset": 0,
                      "utxo_txid": "3c7c0f5c6a0d3f0ab5c8bcef0adf3be56f5aeed8b2dd1504b7a950fc4fee1f46",
                      "utxo_vout": 1
                    },
                    {
                      "inscription_id": "360550a31c9510ed5052c4351619bf68d5ae3f218bf2e9c1092090dbcf86acb3i0",
                      "satoshis": "546",
                      "utxo_block_height": 850976,
                      "utxo_confirmations": 15735,
                      "utxo_sat_offset": 0,
                      "utxo_txid": "3c7c0f5c6a0d3f0ab5c8bcef0adf3be56f5aeed8b2dd1504b7a950fc4fee1f46",
                      "utxo_vout": 1
                    }
                  ],
                  "last_updated": {
                    "block_hash": "1c3e7cd9e46bd6d0adb1ae0b52ec1a1ddfaa0cb61a41a1b1c26f0d1b0f42a06c",
                    "block_height": 5146710
                  },
                  "next_cursor": null
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/assets/inscriptions/{inscription_id}": {
      "get": {
        "tags": [
          "Inscriptions"
        ],
        "summary": "Inscription Info",
        "description": "Information about an inscription. A preview of the content body is given only if its type is\n\"text/plain\". For the whole content, use the complementary endpoint, namely\n`/assets/inscriptions/{inscription_id}/content_body`.",
        "operationId": "inscription_info",
        "parameters": [
          {
            "name": "inscription_id",
            "in": "path",
            "description": "Inscription ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "7d0a2dd897222913d58fc957b0429526117a0a61c964642fe93b077f328ccec1i0"
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimestampedInscriptionInfo"
                },
                "example": {
                  "data": {
                    "content_body_preview": null,
                    "content_length": 3035,
                    "content_type": "text/html;charset=utf-8",
                    "created_at": 842166,
                    "inscription_id": "7d0a2dd897222913d58fc957b0429526117a0a61c964642fe93b077f328ccec1i0",
                    "inscription_num": 234567
                  },
                  "last_updated": {
                    "block_hash": "1c3e7cd9e46bd6d0adb1ae0b52ec1a1ddfaa0cb61a41a1b1c26f0d1b0f42a06c",
                    "block_height": 5146710
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/assets/inscriptions/{inscription_id}/content_body": {
      "get": {
        "tags": [
          "Inscriptions"
        ],
        "summary": "Content by Inscription ID",
        "description": "Paginated response of the content body byte array of an inscription. This endpoint is\ncomplementary to the \"Inscription Information\" endpoint.",
        "operationId": "content_by_inscription_id",
        "parameters": [
          {
            "name": "inscription_id",
            "in": "path",
            "description": "Inscription ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "7d0a2dd897222913d58fc957b0429526117a0a61c964642fe93b077f328ccec1i0"
          },
          {
            "name": "count",
            "in": "query",
            "description": "The max number of bytes per page",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "type": "integer",
                  "default": 100,
                  "minimum": 0
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor string: the offset in the content body. Use the cursor to query for the next page",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedContentBody"
                },
                "example": {
                  "data": {
                    "content_body_page": "AAECAwQFBgcICQoLDA0ODxAREhMUFRYXGBkaGxwdHh8gISIjJCUmJygpKissLS4vMDEyMzQ1Njc4OTo7PD0+P0BBQkNERUZHSElKS0xNTk9QUVJTVFVWV1hZWltcXV5fYGFiYw==",
                    "remaining_bytes": 2935,
                    "total_length": 3035
                  },
                  "last_updated": {
                    "block_hash": "1c3e7cd9e46bd6d0adb1ae0b52ec1a1ddfaa0cb61a41a1b1c26f0d1b0f42a06c",
                    "block_height": 5146710
                  },
                  "next_cursor": "AAAAAAAAAGQ"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentBody": {
        "type": "object",
        "required": [
          "content_body_page",
          "total_length",
          "remaining_bytes"
        ],
        "properties": {
          "content_body_page": {
            "type": "string",
            "description": "Base64-encoded representation of a slice of the inscription content body. All types supported."
          },
          "remaining_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "Number of bytes remaining in the inscription content body.",
            "minimum": 0
          },
          "total_length": {
            "type": "integer",
            "format": "int64",
            "description": "Number of bytes in entire inscription content body.",
            "minimum": 0
          }
        }
      },
...
      "InscriptionByAddress": {
        "type": "object",
        "required": [
          "inscription_id",
          "satoshis",
          "utxo_sat_offset",
          "utxo_txid",
          "utxo_vout",
          "utxo_block_height",
          "utxo_confirmations"
        ],
        "properties": {
          "inscription_id": {
            "type": "string",
            "description": "String representation of the inscription ID, whose first coordinate is the reveal\ntransaction hash, and the second coordinate is the index of inscription in the reveal\ntransaction."
          },
          "satoshis": {
            "type": "string",
            "description": "Total number of satoshis in the UTxO containing the inscription."
          },
          "utxo_block_height": {
            "type": "integer",
            "format": "int64",
            "description": "Block height of the UTxO containing the inscription.",
            "minimum": 0
          },
          "utxo_confirmations": {
            "type": "integer",
            "format": "int64",
            "description": "Number of confirmations of the block where the UTxO containing the inscription was created.",
            "minimum": 0
          },
          "utxo_sat_offset": {
            "type": "integer",
            "format": "int32",
            "description": "Inscribed sat offset in the UTxO containing it.",
            "minimum": 0
          },
          "utxo_txid": {
            "type": "string",
            "description": "Tx ID of the UTxO containing the inscription."
          },
          "utxo_vout": {
            "type": "integer",
            "format": "int32",
            "description": "Output index of the UTxO containing the inscription.",
            "minimum": 0
          }
        }
      },
      "InscriptionInfo": {
        "type": "object",
        "required": [
          "inscription_id",
          "created_at",
          "inscription_num",
          "content_length"
        ],
        "properties": {
          "content_body_preview": {
            "type": "string",
            "description": "Preview of inscription content body raw data. Max: 100 bytes.\nSupported types: \"text/plain\", \"text/plain;charset=utf-8\", \"application/json\".",
            "nullable": true
          },
          "content_length": {
            "type": "integer",
            "format": "int64",
            "description": "Length of entire inscription content body bytes array.",
            "minimum": 0
          },
          "content_type": {
            "type": "string",
            "description": "Type of the content body.",
            "nullable": true
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Block height of the reveal transaction.",
            "minimum": 0
          },
          "inscription_id": {
            "type": "string",
            "description": "String representation of the inscription ID, whose first coordinate is the reveal\ntransaction hash, and the second coordinate is the index of inscription in the reveal\ntransaction."
          },
          "inscription_num": {
            "type": "integer",
            "format": "int64",
            "description": "Global inscription number.",
            "minimum": 0
          }
        }
      },
...
      "LastUpdated": {
        "type": "object",
        "required": [
          "block_hash",
          "block_height"
        ],
        "properties": {
          "block_hash": {
            "type": "string"
          },
          "block_height": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "PaginatedContentBody": {
        "type": "object",
        "required": [
          "data",
          "last_updated"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ContentBody"
          },
          "last_updated": {
            "$ref": "#/components/schemas/LastUpdated"
          },
          "next_cursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "PaginatedInscriptionByAddress": {
        "type": "object",
        "required": [
          "data",
          "last_updated"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InscriptionByAddress"
            }
          },
          "last_updated": {
            "$ref": "#/components/schemas/LastUpdated"
          },
          "next_cursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TimestampedInscriptionInfo": {
        "type": "object",
        "required": [
          "data",
          "last_updated"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/InscriptionInfo"
          },
          "last_updated": {
            "$ref": "#/components/schemas/LastUpdated"
          }
        }
      },
...
    }
  }
}